continuous forward
move forward continuously if there are no obstacle detected. For instance if all sensors do not detect and object within 10cm, then the function will iterate over the if statement and continue moving the robot forward regardless. There is no need to rewrite power allocation to pins if it is already written.